+++ admin/includes/view_all_users.php
@@ -24,12 +24,21 @@
 
   // ========== Delete User ==========
   if(isset($_GET['delete'])){
-    $the_user_id = $_GET['delete'];
 
-    $query = "DELETE FROM users WHERE user_id = {$the_user_id}";
-    $delete_user_query = mysqli_query($connection, $query);
-    // reload page after delete
-    header("Location: users.php");
+    if(isset($_SESSION['user_role'])){
+
+      if($_SESSION['user_role'] == 'admin'){
+
+
+        $the_user_id = mysqli_real_escape_string($connection, $_GET['delete']);
+
+        $query = "DELETE FROM users WHERE user_id = {$the_user_id}";
+        $delete_user_query = mysqli_query($connection, $query);
+        // reload page after delete
+        header("Location: users.php");
+
+      }
+    }
   }
 
 ?>
